Onlymyspace lookup code issue.
it forgot to check for the existence of the timezone value in the line before splitting it, this returned a list index out of range bug. Fix was easy just added an if to fix it and now it works better. 

The pfp code also bugged when pulling from sources other than discord media urls to fix it I just made it split at } as well this fixed everything it seems.